requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
configparser==5.3.0
cryptography==40.0.2
fastapi==0.95.1
h11==0.14.0
idna==3.4
joblib==1.2.0
jwt==1.3.1
numpy==1.24.3
pandas==2.0.1
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
typing-extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
